AP_Mount: viewpro calc_crc len arg fix
Also remove one unnecessary length check
Also correct length check that meant last byte of buffer was unusable

+++ libraries/AP_Mount/AP_Mount_Viewpro.h
@@ -281,7 +281,7 @@ private:
     void process_packet();
 
     // calculate crc of the received message
-    uint8_t calc_crc(const uint8_t *buf, uint32_t length) const;
+    uint8_t calc_crc(const uint8_t *buf, uint8_t length) const;
 
     // get the length and frame count byte (3rd byte of all messages)
     // length is all bytes after the header including CRC
